chore: Move log related code to `dozer-types` and `dozer-log`.
So `dozer-log-js` and `dozer-log-python` only depend on those two crates.
