+++ dozer-core/src/executor/execution_dag.rs
@@ -20,8 +20,7 @@ use daggy::petgraph::{
     visit::{EdgeRef, IntoEdges, IntoEdgesDirected, IntoNodeIdentifiers},
     Direction,
 };
-
-use super::ExecutorOperation;
+use dozer_types::epoch::ExecutorOperation;
 
 pub type SharedRecordWriter = Rc<RefCell<Option<Box<dyn RecordWriter>>>>;
