+++ dozer-log-js/Cargo.toml
@@ -13,10 +13,7 @@ crate-type = ["cdylib"]
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
 [dependencies]
-dozer-api = { path = "../dozer-api" }
-dozer-core = { path = "../dozer-core" }
 dozer-log = { path = "../dozer-log" }
-dozer-orchestrator = { path = "../dozer-orchestrator" }
 dozer-types = { path = "../dozer-types" }
 
 [dependencies.neon]
