+++ dozer-log/src/schemas.rs
@@ -4,23 +4,20 @@ use std::{
     path::{Path, PathBuf},
 };
 
-use dozer_core::{dag_schemas::DagSchemas, DEFAULT_PORT_HANDLE};
-use dozer_sql::pipeline::builder::SchemaSQLContext;
 use dozer_types::{
     models::api_endpoint::{ApiEndpoint, ApiIndex},
+    serde_json,
     types::{Schema, SchemaIdentifier},
 };
 use std::io::Write;
 
-use crate::{errors::OrchestrationError, utils::get_endpoint_schema_path};
+use crate::{errors::SchemaError, get_endpoint_schema_path};
 
 pub fn write_schemas(
-    dag_schemas: &DagSchemas<SchemaSQLContext>,
+    mut schemas: HashMap<String, Schema>,
     pipeline_dir: PathBuf,
     api_endpoints: &[ApiEndpoint],
-) -> Result<HashMap<String, Schema>, OrchestrationError> {
-    let mut schemas = dag_schemas.get_sink_schemas();
-
+) -> Result<HashMap<String, Schema>, SchemaError> {
     for api_endpoint in api_endpoints {
         let schema = schemas
             .get(&api_endpoint.name)
@@ -30,16 +27,16 @@ pub fn write_schemas(
         let schema_path = get_endpoint_schema_path(&pipeline_dir, &api_endpoint.name);
         if let Some(schema_dir) = schema_path.parent() {
             std::fs::create_dir_all(schema_dir)
-                .map_err(|e| OrchestrationError::InternalError(Box::new(e)))?;
+                .map_err(|e| SchemaError::Filesystem(schema_dir.to_path_buf(), e))?;
         }
 
         let mut file = OpenOptions::new()
             .create(true)
             .write(true)
             .open(&schema_path)
-            .map_err(|e| OrchestrationError::InternalError(Box::new(e)))?;
+            .map_err(|e| SchemaError::Filesystem(schema_path.clone(), e))?;
         writeln!(file, "{}", serde_json::to_string(&schema).unwrap())
-            .map_err(|e| OrchestrationError::InternalError(Box::new(e)))?;
+            .map_err(|e| SchemaError::Filesystem(schema_path, e))?;
 
         schemas.insert(api_endpoint.name.clone(), schema);
     }
@@ -47,19 +44,16 @@ pub fn write_schemas(
     Ok(schemas)
 }
 
-pub fn load_schema(pipeline_dir: &Path, endpoint_name: &str) -> Result<Schema, OrchestrationError> {
+pub fn load_schema(pipeline_dir: &Path, endpoint_name: &str) -> Result<Schema, SchemaError> {
     let path = get_endpoint_schema_path(pipeline_dir, endpoint_name);
 
-    let schema_str = std::fs::read_to_string(&path)
-        .map_err(|_| OrchestrationError::SchemasNotInitializedPath(path.clone()))?;
+    let schema_str =
+        std::fs::read_to_string(&path).map_err(|e| SchemaError::Filesystem(path, e))?;
 
-    serde_json::from_str(&schema_str).map_err(|_| OrchestrationError::DeserializeSchema(path))
+    serde_json::from_str(&schema_str).map_err(Into::into)
 }
 
-fn modify_schema(
-    schema: &Schema,
-    api_endpoint: &ApiEndpoint,
-) -> Result<Schema, OrchestrationError> {
+fn modify_schema(schema: &Schema, api_endpoint: &ApiEndpoint) -> Result<Schema, SchemaError> {
     let mut schema = schema.clone();
     // Generated Cache index based on api_index
     let configured_index =
@@ -73,7 +67,7 @@ fn modify_schema(
         (false, true) => configured_index,
         (false, false) => {
             if !upstream_index.eq(&configured_index) {
-                return Err(OrchestrationError::MismatchPrimaryKey {
+                return Err(SchemaError::MismatchPrimaryKey {
                     endpoint_name: api_endpoint.name.clone(),
                     expected: get_field_names(&schema, &upstream_index),
                     actual: get_field_names(&schema, &configured_index),
@@ -85,10 +79,7 @@ fn modify_schema(
 
     schema.primary_index = index;
 
-    schema.identifier = Some(SchemaIdentifier {
-        id: DEFAULT_PORT_HANDLE as u32,
-        version: 1,
-    });
+    schema.identifier = Some(SchemaIdentifier { id: 0, version: 1 });
     Ok(schema)
 }
 
@@ -102,14 +93,14 @@ fn get_field_names(schema: &Schema, indexes: &[usize]) -> Vec<String> {
 fn create_primary_indexes(
     schema: &Schema,
     api_index: &ApiIndex,
-) -> Result<Vec<usize>, OrchestrationError> {
+) -> Result<Vec<usize>, SchemaError> {
     let mut primary_index = Vec::new();
     for name in api_index.primary_key.iter() {
         let idx = schema
             .fields
             .iter()
-            .position(|fd| fd.name == name.clone())
-            .map_or(Err(OrchestrationError::FieldNotFound(name.to_owned())), Ok)?;
+            .position(|fd: &dozer_types::types::FieldDefinition| fd.name == name.clone())
+            .map_or(Err(SchemaError::FieldNotFound(name.to_owned())), Ok)?;
 
         primary_index.push(idx);
     }
