+++ dozer-orchestrator/src/simple/mod.rs
@@ -2,6 +2,3 @@ mod executor;
 pub mod orchestrator;
 pub use orchestrator::SimpleOrchestrator;
 mod helper;
-mod schemas;
-
-pub use schemas::load_schema;
