+++ dozer-orchestrator/src/simple/orchestrator.rs
@@ -1,21 +1,21 @@
 use super::executor::Executor;
-use super::schemas::load_schema;
 use crate::console_helper::get_colored_text;
 use crate::errors::{DeployError, OrchestrationError};
 use crate::pipeline::{LogSinkSettings, PipelineBuilder};
 use crate::shutdown::ShutdownReceiver;
 use crate::simple::helper::validate_config;
-use crate::simple::schemas::write_schemas;
 use crate::utils::{
     get_api_dir, get_api_security_config, get_cache_dir, get_cache_manager_options,
-    get_endpoint_log_path, get_executor_options, get_file_buffer_capacity, get_grpc_config,
-    get_pipeline_dir, get_rest_config,
+    get_executor_options, get_file_buffer_capacity, get_grpc_config, get_pipeline_dir,
+    get_rest_config,
 };
 use crate::{flatten_join_handle, Orchestrator};
 use dozer_api::auth::{Access, Authorizer};
 use dozer_api::generator::protoc::generator::ProtoGenerator;
 use dozer_api::{grpc, rest, CacheEndpoint};
 use dozer_cache::cache::LmdbRwCacheManager;
+use dozer_cache::dozer_log::get_endpoint_log_path;
+use dozer_cache::dozer_log::schemas::{load_schema, write_schemas};
 use dozer_core::app::AppPipeline;
 use dozer_core::dag_schemas::DagSchemas;
 
@@ -296,7 +296,7 @@ impl Orchestrator for SimpleOrchestrator {
 
         // Write schemas to pipeline_dir and generate proto files.
         let schemas = write_schemas(
-            &dag_schemas,
+            dag_schemas.get_sink_schemas(),
             pipeline_home_dir.clone(),
             &self.config.endpoints,
         )?;
