+++ dozer-types/src/lib.rs
@@ -1,9 +1,11 @@
 pub mod borrow;
 pub mod constants;
+pub mod epoch;
 pub mod errors;
 pub mod field_type;
 pub mod helper;
 pub mod ingestion_types;
+pub mod json_types;
 pub mod models;
 pub mod node;
 #[cfg(test)]
